Exercises VIPER refactor

+++ Workout/Home/Exercises/View/ExercisesViewController.swift
@@ -1,22 +1,22 @@
 //
-//  WorkoutsViewController.swift
+//  ExercisesViewController.swift
 //  Workout
 //
-//  Created by Salvador on 7/22/23.
+//  Created by Salvador on 7/26/23.
 //
 
 import UIKit
 import CoreData
 
-protocol WorkoutsPresenterToViewProtocol: NSFetchedResultsControllerDelegate {
-    var presenter: WorkoutsViewToPresenterProtocol { get }
+protocol ExercisesPresenterToViewProtocol: NSFetchedResultsControllerDelegate {
+    var presenter: ExercisesViewToPresenterProtocol! { get }
+    var fetchedResultsControllerDelegate: NSFetchedResultsControllerDelegate { get }
 }
 
-final class WorkoutsViewController: BaseTableViewController {
-    // MARK: - Properties
-    let presenter: WorkoutsViewToPresenterProtocol
+final class ExercisesViewController: BaseTableViewController, NSFetchedResultsControllerDelegate {
+    let presenter: ExercisesViewToPresenterProtocol!
     
-    init(presenter: WorkoutsViewToPresenterProtocol) {
+    init(presenter: ExercisesViewToPresenterProtocol) {
         self.presenter = presenter
         super.init(nibName: nil, bundle: nil)
         
@@ -34,28 +34,25 @@ final class WorkoutsViewController: BaseTableViewController {
         updateContentUnavailableConfiguration()
     }
     
-    func handleNotificationTap(for identifier: String) {
-        presenter.handleNotificationTap(for: identifier)
+    private func configureTableView() {
+        tableView.register(ExerciseTableViewCell.self)
+        tableView.estimatedRowHeight = UITableView.automaticDimension
+        tableView.alwaysBounceVertical = false
+        tableView.dragDelegate = self
     }
     
-    @objc private func addWorkoutTapped() {
-        presenter.addWorkoutTapped()
+    @objc private func addExerciseTapped() {
+        presenter.addExerciseTapped()
     }
-    
+
     private func setupNavigationBar() {
-        navigationItem.title = String(localized: "Workouts")
+        navigationItem.title = presenter.workoutName
         navigationItem.rightBarButtonItem = UIBarButtonItem(title: nil,
                                                             image: UIImage.add,
                                                             target: self,
-                                                            action: #selector(addWorkoutTapped))
+                                                            action: #selector(addExerciseTapped))
         
         navigationController?.navigationBar.isTranslucent = false
-    }
-    
-    private func configureTableView() {
-        tableView.register(WorkoutTableViewCell.self)
-        tableView.estimatedRowHeight = UITableView.automaticDimension
-        tableView.alwaysBounceVertical = false
     }
     
     private func setupViews() {
@@ -66,51 +63,72 @@ final class WorkoutsViewController: BaseTableViewController {
     
     private func showEmptyState() {
         configureEmptyContentUnavailableConfiguration(image: .ellipsis,
-                                                      text: String(localized: "No workouts at the moment"),
+                                                      text: String(localized: "No exercises at the moment"),
                                                       secondaryText: String(localized: "Start adding on the top-right"))
     }
     
     private func updateContentUnavailableConfiguration() {
         UIView.animate(withDuration: 0.25, animations: {
-            self.presenter.workoutsCount == 0 ? self.showEmptyState() : self.clearContentUnavailableConfiguration()
+            self.presenter.exercisesCount == 0 ? self.showEmptyState() : self.clearContentUnavailableConfiguration()
         })
     }
 }
 
 // MARK: - PresenterToViewProtocol
-extension WorkoutsViewController: WorkoutsPresenterToViewProtocol {}
+extension ExercisesViewController: ExercisesPresenterToViewProtocol {
+    var fetchedResultsControllerDelegate: NSFetchedResultsControllerDelegate { self }
+}
 
 // MARK: - UITableViewDelegate
-extension WorkoutsViewController {
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int { presenter.workoutsCount }
+extension ExercisesViewController {
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        presenter.exercisesCount
+    }
     
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: WorkoutTableViewCell.identifier, for: indexPath) as? WorkoutTableViewCell else {
-            WorkoutTableViewCell.assertCellFailure()
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: ExerciseTableViewCell.identifier, for: indexPath) as? ExerciseTableViewCell else {
+            ExerciseTableViewCell.assertCellFailure()
             return UITableViewCell()
         }
         
-        let exercise = presenter.workout(at: indexPath)
+        let exercise = presenter.exercise(at: indexPath)
         cell.configure(with: exercise)
         return cell
+    }
+    
+    override func tableView(_ tableView: UITableView, moveRowAt sourceIndexPath: IndexPath, to destinationIndexPath: IndexPath) {
+        guard sourceIndexPath != destinationIndexPath else { return }
+        
+        presenter.moveRow(at: sourceIndexPath, to: destinationIndexPath)
     }
 }
 
 // MARK: - UITableViewDataSource
-extension WorkoutsViewController {
+extension ExercisesViewController {
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         presenter.didSelectRow(at: indexPath)
     }
     
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        guard editingStyle == .delete else { return }
+        if editingStyle == .delete { presenter.deleteRow(at: indexPath) }
+    }
+}
+
+// MARK: - UITableViewDragDelegate
+extension ExercisesViewController: UITableViewDragDelegate {
+    func tableView(_ tableView: UITableView, itemsForBeginning session: UIDragSession, at indexPath: IndexPath) -> [UIDragItem] {
+        guard let presenter = presenter else {
+            return []
+        }
         
-        presenter.deleteRow(at: indexPath)
+        let dragItem = UIDragItem(itemProvider: NSItemProvider())
+        dragItem.localObject = presenter.exercise(at: indexPath)
+        return [dragItem]
     }
 }
 
 // MARK: - NSFetchedResultsControllerDelegate
-extension WorkoutsViewController: NSFetchedResultsControllerDelegate {
+extension ExercisesViewController {
     func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
         tableView.beginUpdates()
     }
@@ -140,12 +158,13 @@ extension WorkoutsViewController: NSFetchedResultsControllerDelegate {
         case .delete:
             if let indexPath = indexPath {
                 tableView.deleteRows(at: [indexPath], with: .fade)
+                presenter.didDeleteRow(at: indexPath)
                 updateContentUnavailableConfiguration()
             }
         // An update is reported when an object’s state changes, but the changed attributes aren’t part of the sort keys.
         case .update:
-            if let indexPath = indexPath, let cell = tableView.cellForRow(at: indexPath) as? WorkoutTableViewCell {
-                cell.configure(with: presenter.workout(at: indexPath))
+            if let indexPath = indexPath, let cell = tableView.cellForRow(at: indexPath) as? ExerciseTableViewCell {
+                cell.configure(with: presenter.exercise(at: indexPath))
             }
         // A move is reported when the changed attribute on the object is one of the sort descriptors used in the fetch request.
         // An update of the object is assumed in this case, but no separate update message is sent to the delegate.
@@ -153,11 +172,14 @@ extension WorkoutsViewController: NSFetchedResultsControllerDelegate {
             if let indexPath = indexPath {
                 tableView.deleteRows(at: [indexPath], with: .fade)
             }
-
+            
             if let newIndexPath = newIndexPath {
                 tableView.insertRows(at: [newIndexPath], with: .fade)
             }
-        @unknown default:
+            
+            // Moving might also require to update other cells in the tableView
+            tableView.reloadData()
+        default:
             return
         }
     }
